$version/is-debian-alternative marker file
Instead of relying on the rbenv version name to end in "-debian", we now
check for the existence of a is-debian-alternative marker file to
determine if a given rbenv version is a debian alternative.

This has the advantage of allowing a user to symlink a
.rbenv/versions/2.1.8-debian alternative to .rbenv/versions/2.1.8, and
then use existing .ruby-version files that specify 2.1.8.